[PM-1379] rename DeviceCryptoService to DeviceTrustCryptoService - refactors to prevent side effects
