+++ src/Core/Abstractions/IDeviceTrustCryptoService.cs
@@ -3,7 +3,7 @@ using Bit.Core.Models.Domain;
 
 namespace Bit.Core.Abstractions
 {
-    public interface IDeviceCryptoService
+    public interface IDeviceTrustCryptoService
     {
         Task<DeviceResponse> TrustDeviceAsync();
         Task<SymmetricCryptoKey> GetDeviceKeyAsync();
